Move `MastodonClient` to `network..requester.client` package
