+++ core/mastodon/src/main/java/br/com/orcinus/orca/core/mastodon/feed/profile/post/stat/ToggleableStat.extensions.kt
@@ -19,7 +19,7 @@ import br.com.orcinus.orca.core.feed.profile.Profile
 import br.com.orcinus.orca.core.feed.profile.post.stat.toggleable.ToggleableStat
 import br.com.orcinus.orca.core.mastodon.feed.profile.post.MastodonPost
 import br.com.orcinus.orca.core.mastodon.instance.SomeMastodonInstance
-import br.com.orcinus.orca.core.mastodon.network.client.authenticateAndPost
+import br.com.orcinus.orca.core.mastodon.network.requester.client.authenticateAndPost
 import br.com.orcinus.orca.core.module.CoreModule
 import br.com.orcinus.orca.core.module.instanceProvider
 import br.com.orcinus.orca.std.injector.Injector
