+++ core/mastodon/src/main/java/br/com/orcinus/orca/core/mastodon/feed/profile/type/followable/MastodonFollowableProfile.kt
@@ -23,7 +23,7 @@ import br.com.orcinus.orca.core.feed.profile.type.followable.FollowableProfile
 import br.com.orcinus.orca.core.mastodon.feed.profile.MastodonProfile
 import br.com.orcinus.orca.core.mastodon.feed.profile.MastodonProfilePostPaginator
 import br.com.orcinus.orca.core.mastodon.instance.SomeMastodonInstance
-import br.com.orcinus.orca.core.mastodon.network.client.authenticateAndPost
+import br.com.orcinus.orca.core.mastodon.network.requester.client.authenticateAndPost
 import br.com.orcinus.orca.std.image.SomeImageLoader
 import br.com.orcinus.orca.std.injector.Injector
 import br.com.orcinus.orca.std.markdown.Markdown
