+++ core/mastodon/src/test/java/br/com/orcinus/orca/core/mastodon/network/requester/client/ClientResponseProvider.kt
@@ -13,7 +13,7 @@
  * not, see https://www.gnu.org/licenses.
  */
 
-package br.com.orcinus.orca.core.mastodon.network.client
+package br.com.orcinus.orca.core.mastodon.network.requester.client
 
 import io.ktor.client.HttpClient
 import io.ktor.client.engine.mock.MockRequestHandleScope
