+++ core/mastodon/src/test/java/br/com/orcinus/orca/core/mastodon/network/requester/client/TestLogger.extensions.kt
@@ -13,7 +13,7 @@
  * not, see https://www.gnu.org/licenses.
  */
 
-package br.com.orcinus.orca.core.mastodon.network.client
+package br.com.orcinus.orca.core.mastodon.network.requester.client
 
 /** [Logger] returned by [test]. */
 private val testLogger =
